// src/shared/api/generated/model/capsuleDetailResponseOneOf.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Sabujak API
 * 사부작 백엔드 API 문서입니다.
 * OpenAPI spec version: 1.0.0
 */

export type CapsuleDetailResponseOneOf = {
  id: string;
  /**
   * 사용자 노출용 slug 식별자
   * @minLength 1
   * @maxLength 50
   * @pattern ^[a-z0-9]+(?:-[a-z0-9]+)*$
   */
  slug: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  title: string;
  openAt: string;
  expiresAt: string;
  /** 캡슐 수정 optimistic locking 버전 */
  version: number;
  createdAt: string;
  updatedAt: string;
  isOpen: boolean;
  messageCount: number;
};